Verify that predictions are not affected by matrix method
Moving forward, we would like to use the new matrix DWPC
calculation method, because it significantly reduces the runtime.
However, we need to verify that the new DWPC calculation
method does not significantly influence the resulting predictions.

Here we demonstrate that the results from the new technique are
comparable but not exactly equal to the original Neo4j calculation
method.